move version information to separate file for relative imports
