+++ bmds/__init__.py
@@ -3,8 +3,6 @@
 
 isort:skip_file
 """
-__version__ = "1.0.0.dev"
-
 import os
 
 os.environ.setdefault("SIMPLE_SETTINGS", "bmds.settings")  # noqa
@@ -14,3 +12,4 @@ from .datasets import *  # noqa
 from .plotting import *  # noqa
 from .session import *  # noqa
 from .utils import citation  # noqa
+from .version import __version__  # noqa
