+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = bmds
-version = attr: bmds.__version__
+version = attr: bmds.version.__version__
 description = A Python interface for the US EPA Benchmark dose modeling software (BMDS)
 long_description = file: README.rst
 author = Andy Shapiro
